Adding time manipulations functions

+++ srcs/get_rules/get_coder_rules.c
@@ -6,19 +6,20 @@
 /*   By: joesanto <joesanto@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/01/06 21:21:57 by joesanto          #+#    #+#             */
-/*   Updated: 2026/01/06 21:23:19 by joesanto         ###   ########.fr       */
+/*   Updated: 2026/01/08 18:02:29 by joesanto         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "codexion.h"
+#include <stdint.h>
 
-int	get_coder_rules(int *burnout, int *compile, int *debug, int *refactor)
+int	get_coder_rules(uint32_t *burnout, uint32_t *compile, uint32_t *debug, uint32_t *refactor)
 {
-	static int	already_initialized = FALSE;
-	static int	time_to_burnout;
-	static int	time_to_compile;
-	static int	time_to_debug;
-	static int	time_to_refactor;
+	static uint8_t	already_initialized = FALSE;
+	static uint32_t	time_to_burnout;
+	static uint32_t	time_to_compile;
+	static uint32_t	time_to_debug;
+	static uint32_t	time_to_refactor;
 
 	if (already_initialized == FALSE)
 	{
